all problems fixed
