+++ src/components/header/Profile.js
@@ -1,4 +1,4 @@
-import React, { useRef, useState} from "react";
+import React, { useRef, useState, useEffect} from "react";
 import styled from "styled-components";
 import { VscChevronDown, VscChevronUp } from "react-icons/vsc";
 import {useNavigate} from "react-router-dom"
@@ -13,10 +13,16 @@ export default function Profile(){
  
   const {modal, setModal} = useContext(UserContext);
   const {local, setLocal} = useContext(UserContext);
+  const [imgLocal, setImgLocal] = useState('')
 
   useOutsideAlerter(wrapperRef, setModal);
 
   const navigate = useNavigate();
+
+  useEffect(()=>{
+    const img =localStorage.getItem("img");
+    setImgLocal(img);
+  },[])
 
   function HandleClick(){
     let el = document.getElementById('menu').hidden;
@@ -46,7 +52,9 @@ export default function Profile(){
       :
       <IconDown onClick={HandleClick}/> }
 
-      <ProfileImage onClick={HandleClick}></ProfileImage>
+      <ProfileImage onClick={HandleClick}>
+        <img src ={imgLocal} alt =''></img>
+      </ProfileImage>
       
       <Menu id={'menu'} hidden={true}>
         <p onClick={HandleLogout}>Logout</p>
@@ -88,7 +96,6 @@ width: 25%;
 height: 40px;
 margin-top: 15px;
 margin-right: 10px;
-
 `
 
 const IconUp = styled(VscChevronUp)`
@@ -106,5 +113,15 @@ width: 50px;
 height: 50px;
 border-radius: 50%;
 margin-top: 5px;
+display: flex;
+justify-content: center;
+align-items: center;
+
+img{
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+  object-fit: cover;
+}
 
 `
